refactor(lua): reorganize codebase
