+++ lua/cord/api/log/init.lua
@@ -1,7 +1,7 @@
 local logfile = os.getenv 'CORD_LOG_FILE'
 
 if logfile and logfile ~= '' then
-  return require 'cord.internal.log.file'
+  return require 'cord.api.log.file'
 else
-  return require 'cord.internal.log.notify'
+  return require 'cord.api.log.notify'
 end
