+++ lua/cord/internal/activity/workspace.lua
@@ -1,7 +1,7 @@
 local Future = require 'cord.core.async.future'
 local async = require 'cord.core.async'
 local fs = require 'cord.core.uv.fs'
-local config = require 'cord.internal.config'
+local config = require 'cord.api.config'
 local logger = require 'cord.api.log'
 
 local M = {}
